# Mixed analysis: Disregard errors from totals