Disregard errors from mixed analysis and don't require all threads to have completed for the mixed totals.

This could make totals and partial results diverge quite a bit and makes the total operations not match up.

For mixed operations time filter operations consistently by doing only one time segmentation and use that for individual ops as well.

If there are errors don't do time filtering, so we don't hide them.

Before:

```
1354872 operations loaded... Done!
Mixed operations.
Operation: GET, 64%, Concurrency: 96, Duration: 10m0s.
Errors: 44
 * Throughput: 49.27 MiB/s, 1128.98 obj/s

Operation: PUT, 64%, Concurrency: 96, Duration: 10m0s.
Errors: 19
 * Throughput: 49.39 MiB/s, 1128.98 obj/s

Cluster Total: 94.71 MiB/s, 2169.14 obj/s
```

After:

```
1354872 operations loaded... Done!
Mixed operations.
Operation: GET, 50%, Concurrency: 96, Duration: 10m0s.
Errors: 44
 * Throughput: 49.27 MiB/s, 1128.98 obj/s

Operation: PUT, 50%, Concurrency: 96, Duration: 10m0s.
Errors: 19
 * Throughput: 49.39 MiB/s, 1128.98 obj/s

Cluster Total: 98.66 MiB/s, 2257.96 obj/s over 10m0s.
```